+++ example/lib/animate_camera.dart
@@ -60,10 +60,10 @@ class AnimateCameraState extends State<AnimateCamera> {
                         .animateCamera(
                           CameraUpdate.newCameraPosition(
                             const CameraPosition(
-                              bearing: 270.0,
-                              target: LatLng(51.5160895, -0.1294527),
-                              tilt: 30.0,
-                              zoom: 17.0,
+                              bearing: 0.0,
+                              target: LatLng(18.6094114, 103.5862087),
+                              tilt: 0.0,
+                              zoom: 3.0,
                             ),
                           ),
                         )
@@ -77,7 +77,7 @@ class AnimateCameraState extends State<AnimateCamera> {
                     mapController
                         .animateCamera(
                           CameraUpdate.newLatLng(
-                            const LatLng(56.1725505, 10.1850512),
+                            const LatLng(16.4533864, 107.5359136),
                           ),
                           duration: Duration(seconds: 5),
                         )
@@ -106,7 +106,7 @@ class AnimateCameraState extends State<AnimateCamera> {
                   onPressed: () {
                     mapController.animateCamera(
                       CameraUpdate.newLatLngZoom(
-                        const LatLng(37.4231613, -122.087159),
+                        const LatLng(21.0227384, 105.8163641),
                         11.0,
                       ),
                     );
@@ -116,7 +116,7 @@ class AnimateCameraState extends State<AnimateCamera> {
                 TextButton(
                   onPressed: () {
                     mapController.animateCamera(
-                      CameraUpdate.scrollBy(150.0, -225.0),
+                      CameraUpdate.scrollBy(50.0, 25.0),
                     );
                   },
                   child: const Text('scrollBy'),