+++ example/lib/views/example_page.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:vietmap_gl_example/constant.dart';
 
 import 'package:vietmap_flutter_gl/vietmap_flutter_gl.dart';
 
@@ -38,8 +39,7 @@ class _VietmapExampleMapViewState extends State<VietmapExampleMapView> {
           centerTitle: true),
       body: Stack(children: [
         VietmapGL(
-          styleString:
-              "https://maps.vietmap.vn/api/maps/light/styles.json?apikey=YOUR_API_KEY_HERE",
+          styleString: YOUR_STYLE_URL_HERE,
           trackCameraPosition: true,
           onMapCreated: _onMapCreated,
           initialCameraPosition: const CameraPosition(