+++ vietmap_gl_platform_interface/lib/vietmap_gl_platform_interface.dart
@@ -1,6 +1,7 @@
 library vietmap_gl_platform_interface;
 
 import 'dart:async';
+import 'dart:collection';
 import 'dart:convert';
 import 'dart:math';
 import 'package:flutter/foundation.dart';
@@ -8,7 +9,7 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
-import 'package:vietmap_gl_platform_interface/extensions/color_extension.dart';
+
 // part 'src/view_wrappers.dart';
 part 'src/annotation.dart';
 part 'src/callbacks.dart';
@@ -16,9 +17,15 @@ part 'src/camera.dart';
 part 'src/circle.dart';
 part 'src/polyline.dart';
 part 'src/location.dart';
-part 'src/method_channel_mapbox_gl.dart';
+part 'src/method_channel_vietmap_gl.dart';
 part 'src/symbol.dart';
 part 'src/polygon.dart';
 part 'src/ui.dart';
 part 'src/vietmap_gl_platform_interface.dart';
 part 'src/source_properties.dart';
+part 'src/helper/vietmap_math_util.dart';
+part 'src/helper/vietmap_polygon_util.dart';
+part 'src/helper/vietmap_polyline_decode.dart';
+part 'src/helper/vietmap_spherical_util.dart';
+part 'extensions/color_extension.dart';
+part 'src/helper/vietmap_polyline_utils.dart';