+++ CHANGELOG.md
@@ -1,9 +1,36 @@
+## 1.3.4, Dec 18, 2023
+* Upgrade `vietmap_gl_platform_interface` to version `0.1.4`
+* Fix `VietmapPolylineDecoder` `encode` and `decode` function
+* Fix draw polyline below the road name
+## 1.3.3, Dec 15, 2023
+* Provide query `Point` from tile map with `queryRenderedFeatures` function inside `VietmapMapController` class
+## 1.3.2, Dec 11, 2023
+* Provide `nearestLatLngOnLine` and `splitRouteByLatLng` function inside `VietmapPolyline` class, which help show shipper/vehicle location on route and remove route traveled by shipper/vehicle.
+* `splitRouteByLatLng` will split route into 2 parts, which are `routeBeforeLatLng` and `routeAfterLatLng`, you can use one of them to to update the route of shipper/vehicle traveled.
+* Create `VietmapPolyline` class to handle more complex polyline
+## 1.3.1, Dec 05, 2023
+* Add document for `StaticMarkerLayer` and `StaticMarker`
+## 1.3.0, Nov 28, 2023
+* Calculate `StaticMarker` angle when user rotate the map
+## 1.2.7, Nov 27, 2023
+* Fix `StaticMarkerLayer` render 
+## 1.2.5, Nov 23, 2023
+* Fix `StaticMarkerLayer` render
+* Fix bearing anchor display not correct
+* Cập nhật tài liệu `tiếng Việt`
+* Update `README.md` document
+## 1.2.4, Nov 22, 2023
+* Add new `StaticMarkerLayer` and `StaticMarker` to render static marker, which rotate with the map when user rotate the map
+## 1.2.2, Nov 21, 2023
+* Update iOS configuration
+## 1.2.1, Nov 15, 2023
+* Fix android map style
 ## 1.2.0, Nov 15, 2023
-* Update version 2.0.0 of android and ios native map sdk
+* Update version` 2.0.0` of android and ios native map sdk
 ## 1.1.8, Nov 14, 2023
 * Fix white screen error when app to background and back to foreground
 ## 1.1.7, Sep 25, 2023
-* Update README.md document
+* Update `README.md` document
 * Update map style url
 ## 1.1.6, Sep 25, 2023
 * Fix duplicate marker when add new marker to marker layer
@@ -24,22 +51,22 @@
 ## 1.0.8, Aug 14, 2023
 * Fix warning on load drawable image
 ## 1.0.7, Aug 10, 2023
-* Fix MarkerLayer keep state when clear all markers from list
+* Fix `MarkerLayer` keep state when clear all markers from list
 ## 1.0.6, Aug 10, 2023
 * Fix crash app when enable location tracking on android
-* Fix MarkerLayer dispose
+* Fix `MarkerLayer` dispose
 
 ## 1.0.5, Aug 9, 2023
-* Fix onMapRendered callback
+* Fix `onMapRendered` callback
 ## 1.0.4, Aug 9, 2023
 * Fix crash app when enable location tracking on android
-* Add onMapRendered callback
+* Add `onMapRendered` callback
 ## 1.0.3, Aug 8, 2023
 * Fix crash app when update marker location
 ## 1.0.2, Aug 2, 2023
 * Remove unnecessary code
 ## 1.0.1, Aug 2, 2023
-* Fix MarkerLayer render
+* Fix `MarkerLayer` render
 * Allow user to set marker icon align with screen position
 * Allow user to set marker icon size
 ## 1.0.0, Aug 2, 2023